Updated Pressure solvers + minor modifications
Major update of the pressure solver:
  - use of the latest Scheduled Relaxation Jacobi schemes derived in
the literature.
  - changed a bit the calling sequence of the Poisson solver such that
we can easily switch to Algebraic Multigrid Solver in the future if
needed (the fastest Poisson solver available at the moment).

Added a couple of extra subroutines:
  - velocity_bc in tpls_fluids in order to easily apply the velocity
boundary conditions.
  - pressure_bc in tpls_pressure_solver to easily apply the pressure
boundary conditions.
  - etc…

Nothing should change regarding how to use the code however. These
changes are supposed to be completely transparent for regular users.